Add new continuous VI
There is now a new VI where the configuration of the function generator
and the oscope occurs before the oscope starts reading any data. Future
improvements include using a set channel of the oscope (or of the
function generator) to know when the signals are beginning to be sent,
and then to record the wave that the FN generator is sending over as
well as the data the the oscope is reading (via a different channel) to
a spreadsheet, and additionally to a graph in LabView for visualization
purposes.