More experimentation with capturing data
Questions remain with how we will be capturing the data. The fn
generator should be sending a burst every ~1 ms, and this has been
configured. However, the oscilloscope isn't automatically configuring
itself in order to view this wave well. Perhaps the fn generator should
be sending a signal to the oscope to notify it to prepare to capture
right before the wave is actually sent. Continuous capture is not
desired.