Should be less reliant on a LabVIEW loop
Future changes should try to make the LabVIEW code simpler in order
to facilitate understanding of the code and ensuring that the code
does what we expect it to do. Changes in this commit are trying to
get away from the giant loop thing and towards recording a single
burst. There is also a better way of recording data.