Limited functionality present
The oscilloscope is being unusual and recording a very wide
(unchangeable?) amount of data for higher frequency waves (e.g. 1Mhz).
This is resulting in the presentation of the actual wave being poor,
as most of the data is just flatline noise. When operating with lower
frequencies (e.g. 1 kHz) things work somewhat as expected. Future plans
include trying to just use the default VIs that are for the oscilloscope
and manually setting everything else up.